mlx5: share mlx5 PCI device detection
Move PCI detection by IB device from mlx5 PMD to the common code.
