+++ drivers/net/mlx5/mlx5_stats.c
@@ -13,9 +13,12 @@
 #include <rte_common.h>
 #include <rte_malloc.h>
 
+#include <mlx5_common.h>
+
 #include "mlx5_defs.h"
 #include "mlx5.h"
 #include "mlx5_rxtx.h"
+
 
 static const struct mlx5_counter_ctrl mlx5_counters_init[] = {
 	{
